Properly update preview when attributes are updates
Adds a form which renders in the block overlay for ease of use. (For
now, it contains exactly the same controls as in the block inspector.)

Changes to the props in either will trigger a preview fetch, which then
updates the state and triggers a re-render.

Note that the Sandbox component cannot be updated in response to a props
change, so this fudges an update by removing the content and preview
values from the state - so that the component will unmount - and then
sets the state with the updated values so that it can re-render.